Change Interger to Long so no cast will be needed

+++ greenhouse/src/main/java/com/smartgreenhouse/greenhouse/dto/greenhouse/GreenhouseOverviewDTO.java
@@ -16,5 +16,5 @@ public class GreenhouseOverviewDTO {
     private Double currentTemperature;
     private Double currentMoisture;
     private Boolean autoWateringEnabled;
-    private Integer activeSensorCount;
+    private Long activeSensorCount;
 }
